+++ smartcontract/sdk/rs/src/geolocation/geolocation_user/mod.rs
@@ -1,0 +1,8 @@
+pub mod add_target;
+pub mod create;
+pub mod delete;
+pub mod get;
+pub mod list;
+pub mod remove_target;
+pub mod update;
+pub mod update_payment_status;