+++ smartcontract/sdk/rs/src/geolocation/mod.rs
@@ -1,3 +1,4 @@
 pub mod client;
 pub mod geo_probe;
+pub mod geolocation_user;
 pub mod programconfig;